services/browser/requirements.txt
fastapi
uvicorn[standard]
playwright
pydantic

# OpenTelemetry
opentelemetry-api>=1.24.0
opentelemetry-sdk>=1.24.0
opentelemetry-instrumentation-fastapi>=0.45b0
opentelemetry-exporter-otlp-proto-http>=1.24.0
opentelemetry-semantic-conventions>=0.45b0

